testing/device-postmarketos-trailblazer: X1E Slim 7x support
Extend existing X1 Elite support with the modules needed to bring up the
Lenovo Yoga Slim 7x.

We could use the EFI framebuffer in the initramfs, however the
regulators are turned off before we get there so it goes blank. Pull in
all the bits for proper display/panel support.

Signed-off-by: Caleb Connolly <caleb@postmarketos.org>