postmarketos-initramfs: prevent busyloop in syslogd
when stdin is /dev/null (e.g. when console=null is on the kernel
cmdline) syslogd will call read() over and over and never handle the
SIGTERM we send it, pinning a CPU core on every boot.

Provide something on stdin that will always block the read instead, the
signal will wake it up and it will properly exit on SIGTERM.

Signed-off-by: Caleb Connolly <caleb@postmarketos.org>